Added LaneBuilder, for unique lane´s id and in the Lane are some getter changes.
